move Splitter to utils and add accumulate for Python <= 3.3
