+++ src/cclib/parser/qchemparser.py
@@ -18,21 +18,6 @@ from . import logfileparser
 from . import utils
 
 
-class Splitter:
-    def __init__(self, widths):
-        self.start_indices = [0] + list(itertools.accumulate(widths))[:-1]
-        self.end_indices = list(itertools.accumulate(widths))
-
-    def split(self, line):
-        elements = [line[start:end].strip() for (start, end) in zip(self.start_indices, self.end_indices)]
-        for i in range(1, len(elements)):
-            if elements[-1] == '':
-                elements.pop()
-            else:
-                break
-        return elements
-
-
 class QChem(logfileparser.Logfile):
     """A Q-Chem 4 log file."""
 
@@ -232,7 +217,7 @@ class QChem(logfileparser.Logfile):
         line = next(inputfile)
         assert len(line.split()) == min(self.ncolsblock, ncols)
         colcounter = 0
-        split_fixed = Splitter((4, 3, 5, 6, 10, 10, 10, 10, 10, 10))
+        split_fixed = utils.Splitter((4, 3, 5, 6, 10, 10, 10, 10, 10, 10))
         while colcounter < ncols:
             # If the line is just the column header (indices)...
             if line[:5].strip() == '':
